src/main/webapp/app/entities/vrsta-usluge/delete/vrsta-usluge-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { provideHttpClient, HttpResponse } from '@angular/common/http';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { VrstaUslugeService } from '../service/vrsta-usluge.service';

import { VrstaUslugeDeleteDialogComponent } from './vrsta-usluge-delete-dialog.component';

describe('VrstaUsluge Management Delete Component', () => {
  let comp: VrstaUslugeDeleteDialogComponent;
  let fixture: ComponentFixture<VrstaUslugeDeleteDialogComponent>;
  let service: VrstaUslugeService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [VrstaUslugeDeleteDialogComponent],
      providers: [provideHttpClient(), NgbActiveModal],
    })
      .overrideTemplate(VrstaUslugeDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(VrstaUslugeDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(VrstaUslugeService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
